me when i may have broken the canyon or the map just might be weird but i have no idea so uhhh good luck to the next person :)
