+++ Assets/scripts/UI/Selection/SelectionMenuLoadTexts.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a7d0c307bafe7804ca13f593e30e4426
