--- Assets/scripts/UI/Selection/SelectionMenuLoading.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a7d0c307bafe7804ca13f593e30e4426
