+++ Assets/scripts/UI/Selection/SpecialLoadTextBehaviour.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 70ea5af1b5b3fbd439acc385918c4207
