--- Assets/scripts/UI/Selection/loadtextbehaviour.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 789afd6063469e842be6149398743b87
