+++ Assets/terrain/New Terrain 1.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 82f6633a6af73ef47942c3bbc0b0a357
+guid: 1f1f5548fd5b2c148a6f71999bc149cf
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 15600000
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
