--- Assets/terrain/TerrainData_365caccb-70d8-4e03-9bde-1f51b1530741.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 845dda94a166f1c43b18e396e2025306
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
