--- Assets/terrain/TerrainData_58fcba3d-b3cf-43b0-a688-226276526259.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 272955bd3667da044946914d9cfc02f4
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
