+++ Assets/terrain/TerrainData_5fd327af-aca8-4af0-901a-aa135728399c.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 7b9b49eb86ebd7e4bbbb5f81f4721392
+guid: 0b6e70253df7b5e43a07ef84c2cdfbf7
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 15600000
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
