+++ Assets/terrain/TerrainData_6d2877fa-df0f-4b79-99c3-98c6e6b397d9.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: aee5cee49d726d84086ab126e5a16759
+guid: bbafe3926d3899e45886c3f573b985b4
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 15600000
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
