--- Assets/terrain/TerrainData_cb431182-6809-42c2-9271-8122f9ce44f5.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6b6177026118a984682e057e01f11dff
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
