+++ Assets/terrain/scenes/canyon.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 82f6633a6af73ef47942c3bbc0b0a357
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
