+++ Assets/terrain/scenes/canyonExtra1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7b9b49eb86ebd7e4bbbb5f81f4721392
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
