+++ Assets/terrain/scenes/canyonExtra2.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aee5cee49d726d84086ab126e5a16759
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
