+++ Assets/terrain/scenes/canyonExtra3.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 272955bd3667da044946914d9cfc02f4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
