+++ Assets/terrain/scenes/canyonExtra4.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6b6177026118a984682e057e01f11dff
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
