+++ Assets/terrain/scenes/canyonExtra5.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 845dda94a166f1c43b18e396e2025306
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
