--- Assets/textures/tutorial.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fa72766e84608a14aba02ada5adf1915
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
