--- Assets/textures/tutorial/hotel trivago.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2a88bb7bd1ae4a749a5eada396396016
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
