--- Assets/textures/tutorial/tutorialTiles1.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b3128e8b0d4b17d4eb497580f230c40e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
