--- Assets/textures/tutorial/tutorialTiles_layer1.terrainlayer.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 86a4a7c590b22a546af5149f69edf53a
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 8574412962073106934
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
